requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
-e file:.#egg=jenkins_to_github_notify
    # via -r requirements.in
anyio==3.6.1
    # via starlette
atomicwrites==1.4.1
    # via pytest
attrs==21.4.0
    # via pytest
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via
    #   click
    #   pytest
cryptography==37.0.4
    # via pyjwt
distlib==0.3.5
    # via virtualenv
fastapi==0.79.0
    # via -r requirements.in
filelock==3.7.1
    # via virtualenv
github3-py==3.2.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
identify==2.5.2
    # via pre-commit
idna==3.3
    # via
    #   anyio
    #   requests
iniconfig==1.1.1
    # via pytest
mypy==0.971
    # via -r requirements.in
mypy-extensions==0.4.3
    # via mypy
nodeenv==1.7.0
    # via pre-commit
packaging==21.3
    # via pytest
platformdirs==2.5.2
    # via virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via -r requirements.in
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pydantic==1.9.1
    # via fastapi
pyjwt[crypto]==2.4.0
    # via github3-py
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   pytest-datadir
    #   pytest-mock
    #   pytest-regressions
pytest-datadir==1.3.1
    # via pytest-regressions
pytest-mock==3.8.2
    # via -r requirements.in
pytest-regressions==2.3.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via github3-py
python-dotenv==0.20.0
    # via -r requirements.in
pyyaml==6.0
    # via
    #   pre-commit
    #   pytest-regressions
requests==2.28.1
    # via github3-py
six==1.16.0
    # via
    #   python-dateutil
    #   virtualenv
sniffio==1.2.0
    # via anyio
starlette==0.19.1
    # via fastapi
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   mypy
    #   pytest
typing-extensions==4.3.0
    # via
    #   mypy
    #   pydantic
uritemplate==4.1.1
    # via github3-py
urllib3==1.26.10
    # via requests
uvicorn==0.18.2
    # via -r requirements.in
virtualenv==20.15.1
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
